+++ shuup_vendor_reviews/resources.py
@@ -5,8 +5,7 @@
 #
 # This source code is licensed under the OSL-3.0 license found in the
 # LICENSE file in the root directory of this source tree.
-from django.templatetags.static import static
-
+from shuup.core.utils.static import get_shuup_static_url
 from shuup.utils.djangoenv import has_installed
 from shuup.xtheme.resources import add_resource
 
@@ -21,5 +20,13 @@ def add_resources(context, content):
         if match and match.app_name == "shuup_admin":
             return
 
-    add_resource(context, "head_end", "%s?v=0.4.0.css" % static("shuup_vendor_reviews/shuup_vendor_reviews.css"))
-    add_resource(context, "body_end", "%s?v=0.4.0.js" % static("shuup_vendor_reviews/shuup_vendor_reviews.js"))
+    add_resource(
+        context,
+        "head_end",
+        get_shuup_static_url("shuup_vendor_reviews/shuup_vendor_reviews.css", "shuup-vendor-reviews"),
+    )
+    add_resource(
+        context,
+        "body_end",
+        get_shuup_static_url("shuup_vendor_reviews/shuup_vendor_reviews.js", "shuup-vendor-reviews")
+    )