+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Enter here all the changes made to the development version
 
+### Fixed
+
+- Use Shuup static util to fetch the current version of the addon
+
 
 ## [1.0.6] - 2020-10-19